Move waiting for connection outside of locked section
When we are trying to send message while connection is closed
we wait in loop until it is open again. However, because this
is in self._write_lock no other function can access this section
including creating new pika channel. Therefore new channel never
opens and then we get an exception that old channel has already
closed exception.