Catch and log potential exception during the OneShotCallback task
Since __task_inner() is spawned as an asyncio.Task using the channel.closing.add_done_callback(),
the only place where it can be awaited for potential exceptions is under UnderlayChannel.close().
This is done by calling the OneShotCallback.wait() method. However, when the connection to RabbitMQ
is severed, the callback call inside the OneShotCallback task can raise an error when trying to reopen
the connection, and when this happens, the channel instance is replaced with a new channel, which has
a new instance of OneShotCallback.

Because of this, the original OneShotCallback task is never awaited in the UnderlayChannel.close().
So, to properly check for potential exceptions in the original task, we would need to await it somewhere
or call the self.__task.exception() value. This never happens, which produces an error from the asyncio
library: Task exception was never retrieved.

To solve this, we will now catch and log that potential error during the OneShotCallback task execution
which means that we never have to await this task or call self.__task.exception() to retrieve the exception.